Fixes trac:4608: Move r31688 to v1.8 branch (usnic: reduce sizeof(ompi_btl_usnic_addr_t) to 56)
---svn-pre-commit-ignore-below---

r31688 [[BR]]
usnic: reduce sizeof(ompi_btl_usnic_addr_t) to 56 bytes
    
Trivial struct re-ordering to eliminate holes in the middle of the
struct (although there's still a hole at the end) and reduce the
overall size of the struct from 64 to 56 bytes.  Also change mtu from
int to uint16_t; there was no need for it to be that large.

Reviewed by Dave Goodell

cmr=v1.8.2:reviewer=ompi-rm1.8

This commit was SVN r31695.

The following SVN revision numbers were found above:
  r31688 --> open-mpi/ompi@e9c3df652e8af6ac71acf7c7928dbc8605c1e917

The following Trac tickets were found above:
  Ticket 4608 --> https://svn.open-mpi.org/trac/ompi/ticket/4608

+++ ompi/mca/btl/usnic/btl_usnic_proc.c
@@ -588,7 +588,7 @@ static int match_modex(ompi_btl_usnic_module_t *module,
      * the min of the two MTUs?  Another choice is to disqualify this pairing
      * before running the matching algorithm on it. */
     if (*index_out >= 0 &&
-        proc->proc_modex[*index_out].mtu != module->if_mtu) {
+        proc->proc_modex[*index_out].mtu != (uint16_t) module->if_mtu) {
         opal_show_help("help-mpi-btl-usnic.txt", "MTU mismatch",
                     true,
                     ompi_process_info.nodename,
